+++ packages/drivers/src/labwire/drivers/__init__.py
@@ -4,8 +4,17 @@ Example:
     >>> from labwire.drivers import SyringePump
 """
 
+from labwire.drivers._lineproto import LineProtocolClient
 from labwire.drivers.balance import Balance
+from labwire.drivers.endpoints import Endpoint, load_endpoints
 from labwire.drivers.scpi_psu import PowerSupply
 from labwire.drivers.syringe_pump import SyringePump
 
-__all__ = ["Balance", "PowerSupply", "SyringePump"]
+__all__ = [
+    "Balance",
+    "Endpoint",
+    "LineProtocolClient",
+    "PowerSupply",
+    "SyringePump",
+    "load_endpoints",
+]